OmniSwitch: revert recent apps should not change order in quick switcher
Well that was pretty stupid. Changing the order in the
shared list was ofc also changing the order in the
quick switcher. So dont revert global but on access

Thanks to Pablo for the idea.

PS2:
it helps to not only change the positon on fill
but also everytime we need to handle an absolute
position

Change-Id: Ib5c9c34cdcd935da2d3775282bd31356b408709d